Implement priority within semaphore and create a priority lock
