+++ tests/test_zigbee_util.py
@@ -389,186 +389,6 @@ async def test_catching_task_unexpected_exception(to_raise, exception, caplog):
     assert len(records) == 3
 
 
-async def test_dynamic_bounded_semaphore_simple_locking():
-    """Test simple, serial locking/unlocking."""
-    sem = util.DynamicBoundedSemaphore()
-
-    assert "unlocked" not in repr(sem) and "locked" in repr(sem)
-
-    assert sem.value == 0
-    assert sem.max_value == 0
-    assert sem.locked()
-
-    # Invalid max value
-    with pytest.raises(ValueError):
-        sem.max_value = -1
-
-    assert sem.value == 0
-    assert sem.max_value == 0
-    assert sem.locked()
-
-    # Max value is now specified
-    sem.max_value = 1
-    assert not sem.locked()
-    assert sem.max_value == 1
-    assert sem.value == 1
-
-    assert "unlocked" in repr(sem)
-
-    # Semaphore can now be acquired
-    async with sem:
-        assert sem.value == 0
-        assert sem.locked()
-
-    assert not sem.locked()
-    assert sem.max_value == 1
-    assert sem.value == 1
-
-    await sem.acquire()
-    assert sem.value == 0
-    assert sem.locked()
-    sem.release()
-
-    assert not sem.locked()
-    assert sem.max_value == 1
-    assert sem.value == 1
-
-    with pytest.raises(ValueError):
-        sem.release()
-
-
-async def test_dynamic_bounded_semaphore_multiple_locking():
-    """Test multiple locking/unlocking."""
-    sem = util.DynamicBoundedSemaphore(5)
-
-    assert sem.value == 5
-    assert not sem.locked()
-
-    async with sem:
-        assert sem.value == 4
-        assert not sem.locked()
-
-        async with sem, sem, sem:
-            assert sem.value == 1
-            assert not sem.locked()
-
-            with pytest.raises(RuntimeError):
-                async with sem:
-                    assert sem.locked()
-                    assert sem.value == 0
-
-                    raise RuntimeError()
-
-            assert not sem.locked()
-            assert sem.value == 1
-
-        assert sem.value == 4
-        assert not sem.locked()
-
-    assert sem.value == 5
-    assert not sem.locked()
-
-
-async def test_dynamic_bounded_semaphore_runtime_limit_increase(event_loop):
-    """Test changing the max_value at runtime."""
-
-    sem = util.DynamicBoundedSemaphore(2)
-
-    def set_limit(n):
-        sem.max_value = n
-
-    event_loop.call_later(0.1, set_limit, 3)
-
-    async with sem:
-        # Play with the value, testing edge cases
-        sem.max_value = 100
-        assert sem.value == 99
-        assert not sem.locked()
-
-        sem.max_value = 2
-        assert sem.value == 1
-        assert not sem.locked()
-
-        sem.max_value = 1
-        assert sem.value == 0
-        assert sem.locked()
-
-        # Setting it to `0` seems undefined but we keep track of locks so it works
-        sem.max_value = 0
-        assert sem.value == -1
-        assert sem.locked()
-
-        sem.max_value = 2
-        assert sem.value == 1
-        assert not sem.locked()
-
-        async with sem:
-            assert sem.locked()
-            assert sem.value == 0
-            assert sem.max_value == 2
-
-            async with sem:
-                # We're now locked until the limit is increased
-                pass
-
-            assert not sem.locked()
-            assert sem.value == 1
-            assert sem.max_value == 3
-
-        assert sem.value == 2
-        assert sem.max_value == 3
-
-    assert sem.value == 3
-    assert sem.max_value == 3
-
-
-async def test_dynamic_bounded_semaphore_errors(event_loop):
-    """Test semaphore handling errors and cancellation."""
-
-    sem = util.DynamicBoundedSemaphore(1)
-
-    def set_limit(n):
-        sem.max_value = n
-
-    async def acquire():
-        async with sem:
-            await asyncio.sleep(60)
-
-    # The first acquire call will succeed
-    acquire1 = asyncio.create_task(acquire())
-
-    # The remaining two will stall
-    acquire2 = asyncio.create_task(acquire())
-    acquire3 = asyncio.create_task(acquire())
-    await asyncio.sleep(0.1)
-
-    # Cancel the first one, which holds the lock
-    acquire1.cancel()
-
-    # But also cancel the second one, which was waiting
-    acquire2.cancel()
-    with pytest.raises(asyncio.CancelledError):
-        await acquire1
-
-    with pytest.raises(asyncio.CancelledError):
-        await acquire2
-
-    await asyncio.sleep(0.1)
-
-    # The third one will have succeeded
-    assert sem.locked()
-    assert sem.value == 0
-    assert sem.max_value == 1
-
-    acquire3.cancel()
-    with pytest.raises(asyncio.CancelledError):
-        await acquire3
-
-    assert not sem.locked()
-    assert sem.value == 1
-    assert sem.max_value == 1
-
-
 @pytest.mark.parametrize(
     "plot",
     [
