+++ zigpy/util.py
@@ -2,12 +2,10 @@ from __future__ import annotations
 
 import abc
 import asyncio
-import collections
 import functools
 import inspect
 import logging
 import traceback
-import types
 import typing
 import warnings
 
@@ -16,6 +14,7 @@ from cryptography.hazmat.primitives.ciphers import Cipher
 from cryptography.hazmat.primitives.ciphers.algorithms import AES
 from cryptography.hazmat.primitives.ciphers.modes import ECB
 
+from zigpy.datastructures import DynamicBoundedSemaphore  # noqa: F401
 from zigpy.exceptions import ControllerException, ZigbeeException
 import zigpy.types as t
 
@@ -335,120 +334,6 @@ class CatchingTaskMixin(LocalLogMixin):
             self.exception("%s", exc_msg)
 
         return None
-
-
-class DynamicBoundedSemaphore(asyncio.Semaphore):
-    """`asyncio.BoundedSemaphore` with public interface to access and change the max value."""
-
-    def __init__(self, value: int = 0) -> None:
-        self._value: int = value
-        self._max_value: int = value
-
-        self._waiters: collections.deque = collections.deque()
-        self._wakeup_scheduled: bool = False
-
-    @property
-    @functools.lru_cache(maxsize=None)
-    def _loop(self) -> asyncio.BaseEventLoop:
-        return asyncio.get_running_loop()
-
-    def _wake_up_next(self) -> None:
-        while self._waiters:
-            waiter = self._waiters.popleft()
-
-            if not waiter.done():
-                waiter.set_result(None)
-                self._wakeup_scheduled = True
-                return
-
-    @property
-    def value(self) -> int:
-        return self._value
-
-    @property
-    def max_value(self) -> int:
-        return self._max_value
-
-    @max_value.setter
-    def max_value(self, new_value: int) -> None:
-        """Update the semaphore's max value."""
-        if new_value < 0:
-            raise ValueError(f"Semaphore value must be >= 0: {new_value!r}")
-
-        delta = new_value - self._max_value
-        self._value += delta
-        self._max_value += delta
-
-        # Wake up any pending waiters
-        for _ in range(min(len(self._waiters), max(0, delta))):
-            self._wake_up_next()
-
-    @property
-    def num_waiting(self) -> int:
-        return len(self._waiters)
-
-    def locked(self) -> bool:
-        """Returns True if semaphore cannot be acquired immediately."""
-        return self._value <= 0
-
-    async def acquire(self) -> typing.Literal[True]:
-        """Acquire a semaphore.
-
-        If the internal counter is larger than zero on entry, decrement it by one and
-        return True immediately.  If it is zero on entry, block, waiting until some
-        other coroutine has called release() to make it larger than 0, and then return
-        True.
-        """
-
-        # _wakeup_scheduled is set if *another* task is scheduled to wakeup
-        # but its acquire() is not resumed yet
-        while self._wakeup_scheduled or self._value <= 0:
-            fut = self._loop.create_future()
-            self._waiters.append(fut)
-
-            try:
-                await fut
-                # reset _wakeup_scheduled *after* waiting for a future
-                self._wakeup_scheduled = False
-            except asyncio.CancelledError:
-                self._wake_up_next()
-                raise
-
-        assert self._value > 0
-        self._value -= 1
-        return True
-
-    def release(self) -> None:
-        """Release a semaphore, incrementing the internal counter by one.
-
-        When it was zero on entry and another coroutine is waiting for it to become
-        larger than zero again, wake up that coroutine.
-        """
-        if self._value >= self._max_value:
-            raise ValueError("Semaphore released too many times")
-
-        self._value += 1
-        self._wake_up_next()
-
-    async def __aenter__(self) -> None:
-        await self.acquire()
-        return None
-
-    async def __aexit__(
-        self,
-        exc_type: type[BaseException] | None,
-        exc: BaseException | None,
-        traceback: types.TracebackType | None,
-    ) -> None:
-        self.release()
-
-    def __repr__(self) -> str:
-        if self.locked():
-            extra = f"locked, max value:{self._max_value}, waiters:{len(self._waiters)}"
-        else:
-            extra = f"unlocked, value:{self._value}, max value:{self._max_value}"
-
-        return f"<{self.__class__.__name__} [{extra}]>"
 
 
 def deprecated(message: str) -> typing.Callable[[typing.Callable], typing.Callable]:
